refactor. (ddd-strategic-design) 3단계 - 상품 컨텍스트 바운드 분리
