+++ src/main/java/kitchenpos/product/infra/DefaultPurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.product.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
