+++ src/main/java/kitchenpos/product/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.product.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
